backend/requirements-web.txt
# FastAPI Backend Requirements
fastapi==0.115.9
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.1.0
email-validator==2.1.0

# MongoDB
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# Email
emails==0.6
jinja2==3.1.3
aiosmtplib==3.0.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Stripe
stripe==8.5.0

# HTTP Client
httpx>=0.27.0
aiohttp==3.9.3

# Security
cryptography==42.0.2

# Utilities
python-slugify==8.0.4
pillow==10.2.0

# Keep existing project dependencies
requests==2.31.0
beautifulsoup4==4.12.3
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai>=1.68.2,<2.0.0
reportlab==4.1.0
colorlog==6.8.2
